Heatmap Mode
This shows what the heat map looks like on the initial load. In the top left corner you can see the amount of records in the default date range which is the max and min years(2018-2024). This image also showcases what the about tab in the side bar looks like.